+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-05-29 13:50:22
-// Commit: docs(test): update readme and API contracts for test
+// Updated at: 2026-05-29 16:30:53
+// Commit: feat(test): integrate custom telemetry logger in test
 
 class testComponentState {
-  final String commitRef = '1581';
-  final String lastMessage = 'docs(test): update readme and API contracts for test';
-  final DateTime timestamp = DateTime.parse('2026-05-29T13:50:22+05:00');
+  final String commitRef = '1588';
+  final String lastMessage = 'feat(test): integrate custom telemetry logger in test';
+  final DateTime timestamp = DateTime.parse('2026-05-29T16:30:53+05:00');
 }